backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
firebase-admin==6.5.0
supabase==2.18.1
pydantic-settings==2.5.2
httpx==0.28.1
yfinance==0.2.38
pandas==2.2.0
google-generativeai==0.8.3
google-adk==0.5.0
chromadb==0.5.0
textblob==0.18.0
praw==7.7.0
pytrends==4.9.2
sse-starlette==2.1.0
playwright==1.43.0
beautifulsoup4==4.12.0
lxml==5.2.0
tiktoken==0.7.0
numpy==1.26.0
scikit-learn==1.4.0
deprecated==1.2.14
